Update home routes
